requirements.txt
astroid==1.4.8
backports.functools-lru-cache==1.3
configparser==3.5.0
Django==1.10.4
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
drfdocs==0.0.11
haversine==0.4.5
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
Pillow==3.4.2
psycopg2==2.6.2
pylint==1.6.4
six==1.10.0
wrapt==1.10.8
